🔧 Correction du type de conversion de la distance dans les formulaires de création et de mise à jour

+++ src/app/dashboard/runs/[id]/edit/edit-form.tsx
@@ -109,7 +109,7 @@ export function UpdateRunForm({ run }: UpdateRunFormProps) {
                                     placeholder="Entrez la distance"
                                     {...field}
                                     value={field.value === null || field.value === undefined ? "" : field.value}
-                                    onChange={(e) => field.onChange(e.target.value === "" ? null : parseInt(e.target.value))}
+                                    onChange={(e) => field.onChange(e.target.value === "" ? null : parseFloat(e.target.value))}
                                 />
                             </FormControl>
                             <FormMessage />
